+++ EP Mobile/EP Mobile/EPSHcmRiskScore.m
@@ -80,21 +80,21 @@
     return 2;
 }
 
-- (int)numberOfRowsInSection:(int)section {
+- (int)numberOfRowsInSection:(NSInteger)section {
     if (section == 0)
         return 7;
     else
         return 4;
 }
 
-- (int)getOffset:(int)section {
+- (int)getOffset:(NSInteger)section {
     if (section == 1)
         return 7;
     else
         return 0;
 }
 
-- (NSString *)getTitleForHeaderSection:(int)section {
+- (NSString *)getTitleForHeaderSection:(NSInteger)section {
     if (section == 0)
         return @"MAJOR";
     else
