+++ EP Mobile/EP Mobile/EPSQTcCalculatorViewController.m
@@ -142,10 +142,10 @@
 - (IBAction)calculateButtonPressed:(id)sender {
     NSString *input = self.inputField.text;
     NSInteger inputNumber = [input intValue];
-    NSLog(@"The value of inputNumber is %d", inputNumber);
+    NSLog(@"The value of inputNumber is %ld", (long)inputNumber);
     NSString *qt = self.qtField.text;
     NSInteger qtNumber = [qt intValue];
-    NSLog(@"The value of qtNumber is %d", qtNumber);
+    NSLog(@"The value of qtNumber is %ld", (long)qtNumber);
     if (inputNumber <= 0 || qtNumber <= 0) {
         self.resultLabel.textColor = [UIColor darkTextColor];
         self.resultLabel.text = INVALID_ENTRY;
@@ -153,14 +153,14 @@
     }
     if (inputIsRate) {
         inputNumber = 60000.0 / inputNumber;
-        NSLog(@"Converted to RR interval in msec is %d", inputNumber);
+        NSLog(@"Converted to RR interval in msec is %ld", (long)inputNumber);
     }
     NSInteger row = [formulaPicker selectedRowInComponent:0];
     NSString *formula = [formulaData objectAtIndex:row];
     NSLog(@"Formula is %@", formula);
-    NSLog(@"Row is %d", row);
+    NSLog(@"Row is %ld", (long)row);
     NSInteger qtc = [self qtcFromQtInMsec:qtNumber AndIntervalInMsec:inputNumber UsingFormula:row];
-    NSLog(@"QTc = %d", qtc);
+    NSLog(@"QTc = %ld", (long)qtc);
     if (qtc == 0.0) {
         self.resultLabel.textColor = [UIColor darkTextColor];
         self.resultLabel.text = INVALID_ENTRY;
@@ -170,7 +170,7 @@
             self.resultLabel.textColor = [UIColor redColor];
         else
             self.resultLabel.textColor = [UIColor darkTextColor];
-        self.resultLabel.text = [[NSString alloc] initWithFormat:@"QTc is %i msec (%@ formula)", qtc, formula];
+        self.resultLabel.text = [[NSString alloc] initWithFormat:@"QTc is %li msec (%@ formula)", (long)qtc, formula];
         // result text color
     }
     //self.resultLabel.text = resultString;
@@ -219,7 +219,7 @@
     [self setInputType:[sender selectedSegmentIndex]];
 }
 
-- (void)setInputType:(int)index {
+- (void)setInputType:(NSInteger)index {
     if ((inputIsRate = index == RATE_INDEX))
         self.inputField.placeholder = @"Heart Rate (bpm)";
     else
